+++ CHANGES.md
@@ -14,6 +14,7 @@ Changes in 11.6
 - Use vertically align list data option 'valign-middle', program wide
 - CSS add .has-input class to list having input (input or select fields) in ListOutput.fnc.php
 - CSS add .list-save class to Export list link in ListOutput.fnc.php
+- CSS add .list-add-row class (see `$link['add']` parameter) in ListOutput.fnc.php
 - ExplodeDate() use strtotime() in Date.php
 - Remove MonthNWSwitch(), `__mnwswitch_num2char()` & `__mnwswitch_char2num()` functions in Date.php
 - Send wkhtmltopdf error by email in PDF.php